+++ src/descriptor/error.rs
@@ -27,26 +27,19 @@
 /// Errors related to the parsing and usage of descriptors
 #[derive(Debug)]
 pub enum Error {
-    //InternalError,
-    //InvalidPrefix(Vec<u8>),
-    //HardenedDerivationOnXpub,
-    //MalformedInput,
     /// Invalid HD Key path, such as having a wildcard but a length != 1
     InvalidHDKeyPath,
+    /// The provided descriptor doesn't match its checksum
+    InvalidDescriptorChecksum,
 
-    //KeyParsingError(String),
     /// Error thrown while working with [`keys`](crate::keys)
     Key(crate::keys::KeyError),
     /// Error while extracting and manipulating policies
     Policy(crate::descriptor::policy::PolicyError),
 
-    //InputIndexDoesntExist,
-    //MissingPublicKey,
-    //MissingDetails,
     /// Invalid character found in the descriptor checksum
     InvalidDescriptorCharacter(char),
 
-    //CantDeriveWithMiniscript,
     /// BIP32 error
     BIP32(bitcoin::util::bip32::Error),
     /// Error during base58 decoding